--- webapp/src/main/java/org/opentestsystem/rdw/reporting/model/Translation.java
@@ -1,60 +1,0 @@
-package org.opentestsystem.rdw.reporting.model;
-
-import java.util.Locale;
-
-public class Translation {
-
-    private final String code;
-    private final Locale locale;
-    private final String message;
-
-    private Translation(final String code, final Locale locale, final String message) {
-        this.code = code;
-        this.locale = locale;
-        this.message = message;
-    }
-
-    public Locale getLocale() {
-        return locale;
-    }
-
-    public String getCode() {
-        return code;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public static Builder builder() {
-        return new Builder();
-    }
-
-    /**
-     * The builder for the {@link Translation}
-     */
-    public static class Builder {
-        private String code;
-        private Locale locale;
-        private String message;
-
-        public Translation build() {
-            return new Translation(code, locale, message);
-        }
-
-        public Builder withCode(final String code) {
-            this.code = code;
-            return this;
-        }
-
-        public Builder withLocale(final Locale locale) {
-            this.locale = locale;
-            return this;
-        }
-
-        public Builder withMessage(final String message) {
-            this.message = message;
-            return this;
-        }
-    }
-}
