--- webapp/src/main/java/org/opentestsystem/rdw/reporting/repository/GroupGrantRepository.java
@@ -1,17 +1,0 @@
-package org.opentestsystem.rdw.reporting.repository;
-
-import org.opentestsystem.rdw.reporting.model.GroupGrant;
-
-import java.util.List;
-
-public interface GroupGrantRepository {
-
-    /**
-     * Find all {@GroupMembership}s for the given user name
-     *
-     * @param username the user's name
-     * @return collection of {@link org.opentestsystem.rdw.reporting.model.GroupGrant}
-     */
-    List<GroupGrant> findAllForUsername(String username);
-
-}
