+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/security/GroupGrant.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.model;
+package org.opentestsystem.rdw.reporting.security;
 
 import javax.validation.constraints.NotNull;
