+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/security/Permission.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.model;
+package org.opentestsystem.rdw.reporting.security;
 
 import com.google.common.base.MoreObjects;
