+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/security/User.java
@@ -2,8 +2,6 @@ package org.opentestsystem.rdw.reporting.security;
 
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
-import org.opentestsystem.rdw.reporting.model.GroupGrant;
-import org.opentestsystem.rdw.reporting.model.Permission;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
