+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/security/auth/AuthorizationService.java
@@ -1,7 +1,7 @@
 package org.opentestsystem.rdw.reporting.security.auth;
 
-import org.opentestsystem.rdw.reporting.model.GroupGrant;
-import org.opentestsystem.rdw.reporting.model.Permission;
+import org.opentestsystem.rdw.reporting.security.GroupGrant;
+import org.opentestsystem.rdw.reporting.security.Permission;
 import org.opentestsystem.rdw.security.Grant;
 import org.springframework.security.core.GrantedAuthority;
