+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/security/auth/DefaultAuthorizationService.java
@@ -4,11 +4,10 @@ import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Multimap;
 import com.google.common.collect.Multimaps;
-import org.opentestsystem.rdw.reporting.model.GroupGrant;
-import org.opentestsystem.rdw.reporting.model.Permission;
-import org.opentestsystem.rdw.reporting.model.PermissionScope;
+import org.opentestsystem.rdw.reporting.security.GroupGrant;
+import org.opentestsystem.rdw.reporting.security.Permission;
+import org.opentestsystem.rdw.reporting.security.PermissionScope;
 import org.opentestsystem.rdw.reporting.repository.DistrictRepository;
-import org.opentestsystem.rdw.reporting.repository.GroupGrantRepository;
 import org.opentestsystem.rdw.reporting.repository.InstitutionRepository;
 import org.opentestsystem.rdw.reporting.security.Authorities;
 import org.opentestsystem.rdw.security.Grant;
@@ -52,7 +51,7 @@ public class DefaultAuthorizationService implements AuthorizationService {
         this.groupMembershipRepository = groupMembershipRepository;
     }
 
-    public Set<Grant> getGrants(String[] encodedGrants) {
+    public Set<Grant> getGrants(final String[] encodedGrants) {
         final Set<Grant> grants = newLinkedHashSet();
         for (String encodedGrant : encodedGrants) {
             final Grant grant = Grants.from(encodedGrant);
@@ -65,7 +64,7 @@ public class DefaultAuthorizationService implements AuthorizationService {
         return ImmutableSet.copyOf(grants);
     }
 
-    public Set<Permission> getPermissions(Collection<? extends Grant> grants) {
+    public Set<Permission> getPermissions(final Collection<? extends Grant> grants) {
 
         // Groups grants by role to later combine their districts and institutions into a single scope
         // This is to cover the case of multiple non-intersecting grants being given
@@ -108,7 +107,7 @@ public class DefaultAuthorizationService implements AuthorizationService {
         return ImmutableSet.copyOf(permissions);
     }
 
-    public Set<GrantedAuthority> getAuthorities(Collection<? extends Grant> grants) {
+    public Set<GrantedAuthority> getAuthorities(final Collection<? extends Grant> grants) {
         final Map<String, Collection<String>> permissionsByRole = getPermissionsByRole();
         final ImmutableSet.Builder<GrantedAuthority> authorities = ImmutableSet.builder();
         for (Grant grant : grants) {
@@ -127,7 +126,7 @@ public class DefaultAuthorizationService implements AuthorizationService {
         return authorities.build();
     }
 
-    public Map<Long, GroupGrant> getGroupsById(Collection<? extends Permission> permissions, String username) {
+    public Map<Long, GroupGrant> getGroupsById(final Collection<? extends Permission> permissions, final String username) {
         // only lookup groups for those with the group PII permission
         if (permissions.stream().noneMatch(permission -> permission.getId().equals("GROUP_PII_READ"))) {
             return ImmutableMap.of();
@@ -144,7 +143,10 @@ public class DefaultAuthorizationService implements AuthorizationService {
         }
     }
 
-    private PermissionScope createScope(Collection<Grant> grantsWithSameRole, Map<String, Long> districtIdByNaturalId, Map<String, Long> institutionIdByNaturalId) {
+    private PermissionScope createScope(
+            final Collection<Grant> grantsWithSameRole,
+            final Map<String, Long> districtIdByNaturalId,
+            final Map<String, Long> institutionIdByNaturalId) {
 
         if (grantsWithSameRole.size() == 1) {
             return createScope(
@@ -171,7 +173,11 @@ public class DefaultAuthorizationService implements AuthorizationService {
         return new PermissionScope(districtIds.build(), institutionIds.build());
     }
 
-    private PermissionScope createScope(Grant grant, Map<String, Long> districtIdByNaturalId, Map<String, Long> institutionIdByNaturalId) {
+    private PermissionScope createScope(
+            final Grant grant,
+            final Map<String, Long> districtIdByNaturalId,
+            final Map<String, Long> institutionIdByNaturalId) {
+
         if (grant.getEntityLevel() == Grant.EntityLevel.STATE) {
             return PermissionScope.STATEWIDE;
         }
