+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/security/auth/GroupGrantRepository.java
@@ -1,0 +1,17 @@
+package org.opentestsystem.rdw.reporting.security.auth;
+
+import org.opentestsystem.rdw.reporting.security.GroupGrant;
+
+import java.util.List;
+
+public interface GroupGrantRepository {
+
+    /**
+     * Find all {@link GroupGrant}s for the given user name
+     *
+     * @param username the user's name
+     * @return collection of {@link GroupGrant}
+     */
+    List<GroupGrant> findAllForUsername(String username);
+
+}
