+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/security/auth/JdbcGroupGrantRepository.java
@@ -1,19 +1,17 @@
-package org.opentestsystem.rdw.reporting.repository.impl;
+package org.opentestsystem.rdw.reporting.security.auth;
 
-import org.opentestsystem.rdw.reporting.model.GroupGrant;
-import org.opentestsystem.rdw.reporting.repository.GroupGrantRepository;
+import org.opentestsystem.rdw.reporting.security.GroupGrant;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 
+import javax.validation.constraints.NotNull;
 import java.util.List;
 
-import static com.google.common.base.Preconditions.checkNotNull;
-
 @Repository
-public class GroupGrantRepositoryImpl implements GroupGrantRepository {
+public class JdbcGroupGrantRepository implements GroupGrantRepository {
 
     private final NamedParameterJdbcTemplate jdbcTemplate;
 
@@ -21,12 +19,12 @@ public class GroupGrantRepositoryImpl implements GroupGrantRepository {
     private String findAllForUsernameQuery;
 
     @Autowired
-    public GroupGrantRepositoryImpl(final NamedParameterJdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = checkNotNull(jdbcTemplate, "jdbcTemplate must not be null");
+    public JdbcGroupGrantRepository(final @NotNull NamedParameterJdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
     }
 
     @Override
-    public List<GroupGrant> findAllForUsername(String username) {
+    public List<GroupGrant> findAllForUsername(final String username) {
         return jdbcTemplate.query(findAllForUsernameQuery,
                 new MapSqlParameterSource("user_login", username),
                 (results, index) -> new GroupGrant(
@@ -35,4 +33,5 @@ public class GroupGrantRepositoryImpl implements GroupGrantRepository {
                 )
         );
     }
+
 }
