+++ webapp/src/test/java/org/opentestsystem/rdw/reporting/security/auth/DefaultAuthorizationServiceTest.java
@@ -5,11 +5,10 @@ import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import org.junit.Before;
 import org.junit.Test;
-import org.opentestsystem.rdw.reporting.model.GroupGrant;
-import org.opentestsystem.rdw.reporting.model.Permission;
-import org.opentestsystem.rdw.reporting.model.PermissionScope;
+import org.opentestsystem.rdw.reporting.security.GroupGrant;
+import org.opentestsystem.rdw.reporting.security.Permission;
+import org.opentestsystem.rdw.reporting.security.PermissionScope;
 import org.opentestsystem.rdw.reporting.repository.DistrictRepository;
-import org.opentestsystem.rdw.reporting.repository.GroupGrantRepository;
 import org.opentestsystem.rdw.reporting.repository.InstitutionRepository;
 import org.opentestsystem.rdw.security.Grant;
 import org.opentestsystem.rdw.security.service.PermissionService;
